# src/client/generated/w3s-openapi.yaml
openapi: 3.0.3
info:
  title: W3S
  description: All W3S APIs
  version: '1.0'
servers:
  - url: https://api.circle.com/v1/w3s
security:
  - BearerAuth: []
...
paths:
...
  /users:
    post:
      tags:
        - Users and Pins
      summary: Create a user
      description: Create a user.
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
        description: Request body
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
          description: Success response
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        default:
          $ref: '#/components/responses/Default'
    get:
      tags:
        - Users and Pins
      summary: List users
      description: Get all the users under the entity, sorted by the create date DESC order.
      operationId: listUsers
      parameters:
        - $ref: '#/components/parameters/PinStatus'
        - $ref: '#/components/parameters/SecurityQuestionStatus'
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
        - $ref: '#/components/parameters/PageBefore'
        - $ref: '#/components/parameters/PageAfter'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ExtGetAllEndUsersResponse'
          description: Users retrieved
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            Link:
              description: 'Pagination cursor information, it can have four links depending on the page: self, first, next and prev. format: <https://api.circle.com/w3s/users?to=2020-06-02T12:01:00Z&pageSize=50>; rel=\"next\'
              schema:
                type: string
        default:
          $ref: '#/components/responses/Default'
  /users/{id}:
    get:
      summary: Get a user
      tags:
        - Users and Pins
      description: Get user by ID.
      operationId: getUser
      parameters:
        - name: id
          description: User ID
          in: path
          required: true
          schema:
            format: uuid
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/GetEndUserByIDResponse'
          description: Success response
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        default:
          $ref: '#/components/responses/Default'
  /users/token:
    post:
      tags:
        - Users and Pins
      summary: Create a user token
      description: Generate user session and SDK secret key.
      operationId: getUserToken
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateUserTokenRequest'
        description: Request body
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/GenerateUserTokenResponse'
          description: Success response
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        default:
          $ref: '#/components/responses/Default'
  /user:
    get:
      tags:
        - Users and Pins
      summary: Get user status
      description: Get user's PIN and Security Question statuses.
      operationId: getUserByToken
      parameters:
        - $ref: '#/components/parameters/XUserTokenRequired'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/EndUser'
          description: Success response
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        default:
          $ref: '#/components/responses/Default'
  /user/initialize:
    post:
      tags:
        - Users and Pins
      summary: Create a challenge for PIN setup and create wallet(s)
      description: Create a challenge for PIN setup and create wallet(s)
      operationId: createUserWithPinChallenge
      parameters:
        - $ref: '#/components/parameters/XUserTokenRequired'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetPinAndInitWalletRequest'
        description: Request body
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/SetPinResponse'
          description: Success response
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        default:
          $ref: '#/components/responses/Default'
  /user/challenges:
    get:
      tags:
        - Users and Pins
      summary: List challenges
      description: List all challenges by status for a user.
      operationId: listUserChallenges
      parameters:
        - $ref: '#/components/parameters/XUserTokenRequired'
        - $ref: '#/components/parameters/Status'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/GetEndUserChallengesResponse'
          description: Success response
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        default:
          $ref: '#/components/responses/Default'
  /user/challenges/{id}:
    get:
      tags:
        - Users and Pins
      summary: Get a challenge
      description: Get the status of a challenge.
      operationId: getUserChallenge
      parameters:
        - $ref: '#/components/parameters/XUserTokenRequired'
        - name: id
          description: ExtChallenge id to query
          in: path
          required: true
          schema:
            format: uuid
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/GetEndUserChallengeByIDResponse'
          description: Success response
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        default:
          $ref: '#/components/responses/Default'
  /user/pin:
    post:
      tags:
        - Users and Pins
      summary: Create a challenge for PIN setup
      description: Create a challenge for PIN setup without setting up wallets.
      operationId: createUserPinChallenge
      parameters:
        - $ref: '#/components/parameters/XUserTokenRequired'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetPinRequest'
        description: Request body
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/SetPinResponse'
          description: Success response
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        default:
          $ref: '#/components/responses/Default'
    put:
      tags:
        - Users and Pins
      summary: Create a challenge to update PIN
      description: Create a challenge to update a user's PIN via existing PIN.
      operationId: updateUserPinChallenge
      parameters:
        - $ref: '#/components/parameters/XUserTokenRequired'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePinRequest'
        description: Request body
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ChangePinResponse'
          description: Success response
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        default:
          $ref: '#/components/responses/Default'
  /user/pin/restore:
    post:
      tags:
        - Users and Pins
      summary: Create a challenge for PIN restore
      description: Create a challenge to change a user's PIN via Security Questions.
      operationId: createUserPinRestoreChallenge
      parameters:
        - $ref: '#/components/parameters/XUserTokenRequired'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestorePinRequest'
        description: Request body
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/RestorePinResponse'
          description: Success response
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        default:
          $ref: '#/components/responses/Default'
...
  /contracts:
    get:
      tags:
        - View / Update
      summary: List contracts
      description: List all contracts
      operationId: listContracts
      parameters:
        - name: blockchain
          description: Blockchain is the blockchain filter for contracts
          in: query
          schema:
            type: string
            enum:
              - ETH
              - ETH-GOERLI
              - AVAX
              - AVAX-FUJI
              - MATIC
              - MATIC-MUMBAI
        - name: contractInputType
          description: ContractInputType is the input type filter for contracts
          in: query
          schema:
            type: string
            enum:
              - IMPORT
              - BYTECODE
        - name: deployerAddress
          description: DeployerAddress is the deployer wallet address filter for contracts
          in: query
          schema:
            type: string
        - name: name
          description: Name is the name filter for contracts
          in: query
          schema:
            type: string
        - name: status
          description: Status is the status filter for contracts
          in: query
          schema:
            type: string
            enum:
              - PENDING
              - FAILED
              - COMPLETE
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
        - $ref: '#/components/parameters/PageBefore'
        - $ref: '#/components/parameters/PageAfter'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/ContractsResponse'
          description: Contract list retrieved
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            Link:
              description: 'Pagination cursor information. format: <https://api.circle.com/v1/contracts?to=2020-06-02T12:01:00Z&pageSize=50>; rel=\"next\'
              schema:
                type: string
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error response
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
  /contracts/{id}:
    get:
      tags:
        - View / Update
      summary: Get a contract
      description: Get a contract by id. This does not support getting a contract by address.
      operationId: getContract
      parameters:
        - description: Contract id
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/ContractResponse'
          description: Contract retrieved
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error response
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
    patch:
      tags:
        - View / Update
      summary: Update a contract
      description: Update off-chain contract properties
      operationId: updateContract
      parameters:
        - description: Contract id
          in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchContractRequest'
        description: Update contract properties request
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/ContractResponse'
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error response
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
components:
  schemas:
...
    Contract:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/schemas_Id'
        deployerWalletID:
          type: string
          format: uuid
          description: The `id` of the Programmable Wallet that deployed this contract.
          example: f39e3dbd-84af-4d3d-b5ac-98fbe047ce6a
        deploymentTransactionId:
          type: string
          format: uuid
          description: The id of the deployment transaction for this contract.
          example: 00ca46f3-c31b-4e3a-92a9-068d4ff26a63
        implementationContractId:
          type: string
          format: uuid
          description: If this contract is a proxy, the contract ID of the implementation.
          example: f3d81192-1f5d-4d17-ae85-5801a15ea273
        txHash:
          type: string
          description: The string value of the transaction hash.
          example: '0x4a25cc5e661d8504b59c5f38ba93f010e8518966f00e2ceda7955c4b8621357d'
        abiJson:
          $ref: '#/components/schemas/AbiJson'
        archived:
          $ref: '#/components/schemas/Archived'
        blockchain:
          $ref: '#/components/schemas/Blockchain'
        bytecode:
          $ref: '#/components/schemas/Bytecode'
        contractAddress:
          type: string
          description: The on-chain address of this contract.
          example: '0x1e124d7384cd34448ea5907bd0052a79355ab5eb'
        contractInputType:
          $ref: '#/components/schemas/modelContractInputType'
        deployerAddress:
          type: string
          description: The address that created this contract, if deployed.
          example: '0x1bf9ad0cc2ad298c69a2995aa806ee832788218c'
        name:
          type: string
          description: The contract's name.
          example: First Contract
        description:
          $ref: '#/components/schemas/ContractDescription'
        status:
          $ref: '#/components/schemas/modelStatus'
        verificationStatus:
          $ref: '#/components/schemas/modelVerificationStatus'
        metadataLink:
          type: string
          description: The ipfs metadata link of the contract.
          example: https://ipfs.io/ipfs/Qme7ss3ARVgxv6rXqVPiikMJ8u2NLgmgszg13pYrDKEoiu
        updateDate:
          $ref: '#/components/schemas/UpdateDate'
        createDate:
          $ref: '#/components/schemas/schemas_CreateDate'
        sourceCode:
          type: array
          readOnly: true
          items:
            $ref: '#/components/schemas/modelSolFile'
        functions:
          type: array
          description: Functions supported by this contract. Parsed from abi_json.
          items:
            $ref: '#/components/schemas/modelFunction'
        events:
          type: array
          description: Events this contract can emit. Parsed from abi_json.
          items:
            $ref: '#/components/schemas/modelEvent'
    ContractResponse:
      type: object
      properties:
        contract:
          $ref: '#/components/schemas/Contract'
    ContractsResponse:
      type: object
      properties:
        contracts:
          type: array
          description: Contracts are all the contracts that match criteria.
          items:
            $ref: '#/components/schemas/Contract'
...
    PatchContractRequest:
      type: object
      properties:
        name:
          $ref: '#/components/schemas/ContractName'
        description:
          $ref: '#/components/schemas/ContractDescription'
        archived:
          $ref: '#/components/schemas/Archived'
...
    ContractName:
      type: string
      description: The contract's name.
      example: First Contract
    ContractDescription:
      type: string
      description: The description for a contract.
      example: My first hello world contract
    AbiJson:
      type: string
      description: The contract's ABI in a JSON stringified format.
      example: "[\n\t{\n\t\t\"inputs\": [],\n\t\t\"stateMutability\": \"nonpayable\",\n\t\t\"type\": \"constructor\"\n\t},\n\t..."
...
    Archived:
      description: Archived is the updated archived setting of the contract. If true, the contract will not be visible in your dashboard.
      type: boolean
      example: false
...
    Error:
      type: object
      properties:
        code:
          type: integer
          description: Code is the error code
        message:
          type: string
          description: Message is the error message
...
    Blockchain:
      type: string
      enum:
        - ETH-GOERLI
        - ETH
        - AVAX-FUJI
        - AVAX
        - MATIC-MUMBAI
        - MATIC
...
    schemas_Id:
      type: string
      format: uuid
      description: Unique system generated identifier of the resource.
...
    schemas_CreateDate:
      type: string
      format: date-time
      description: Date and time the resource was created. ISO-8601 UTC date/time.
      example: '2023-01-01T12:04:05Z'
...
    schemas_IdempotencyKey:
      type: string
      description: Universally unique identifier (UUID v4) idempotency key. This key is utilized to ensure exactly-once execution of mutating requests. To create a UUIDv4 go to [uuidgenerator.net](https://www.uuidgenerator.net).
      example: a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11
      format: uuid
...
    ExternalUserId:
      type: string
      description: Unique system generated identifier for the user.
      example: ext_user_id_1
      maxLength: 50
      minLength: 5
...
    WalletReferenceId:
      type: string
      description: Optional reference or description used to identify the wallet.
      example: wallet123
...
    PinStatus:
      type: string
      enum:
        - ENABLED
        - UNSET
        - LOCKED
      description: Status of the user PIN code. PIN must be in ENABLED state to authorize requests.
      example: ENABLED
    EndUserStatus:
      type: string
      enum:
        - ENABLED
        - DISABLED
      example: ENABLED
    PinSecurityDetails:
      type: object
      properties:
        failedAttempts:
          type: integer
          description: |
            FailedAttempts the number of times a user has entered an invalid
            pin code.
        lockedDate:
          format: date-time
          type: string
          description: The date and time when the pin code has been locked.
          example: '2023-01-01T12:04:05Z'
        lockedExpiryDate:
          type: string
          format: date-time
          description: The date and time when the pin code will be unlocked.
          example: '2023-01-01T12:04:05Z'
        lastLockOverrideDate:
          type: string
          format: date-time
          description: This is the last date and time when the lock was unlocked by admin manually.
          example: '2023-01-01T12:04:05Z'
    EndUser:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/schemas_Id'
        createDate:
          $ref: '#/components/schemas/schemas_CreateDate'
        pinStatus:
          $ref: '#/components/schemas/PinStatus'
        status:
          $ref: '#/components/schemas/EndUserStatus'
        securityQuestionStatus:
          type: string
          enum:
            - ENABLED
            - UNSET
            - LOCKED
          example: ENABLED
        pinDetails:
          allOf:
            - $ref: '#/components/schemas/PinSecurityDetails'
        securityQuestionDetails:
          allOf:
            - $ref: '#/components/schemas/PinSecurityDetails'
    ExtGetAllEndUsersResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/EndUser'
    CreateUserRequest:
      type: object
      required:
        - userId
      properties:
        userId:
          $ref: '#/components/schemas/ExternalUserId'
    GetEndUserByIDResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/EndUser'
    GenerateUserTokenRequest:
      type: object
      required:
        - userId
      properties:
        userId:
          $ref: '#/components/schemas/ExternalUserId'
    GenerateUserTokenResponse:
      type: object
      required:
        - userToken
      properties:
        userToken:
          type: string
          description: Unique system generated JWT session token. The token will expire after 60 minutes.
        encryptionKey:
          type: string
          description: Encryption key to be used with the SDK for initiating user challenges.
    MetadataField:
      type: object
      properties:
        name:
          type: string
          description: Name or description associated with the wallet.
          example: My walletResponse
        refId:
          $ref: '#/components/schemas/WalletReferenceId'
    SetPinAndInitWalletRequest:
      type: object
      required:
        - idempotencyKey
      properties:
        idempotencyKey:
          $ref: '#/components/schemas/schemas_IdempotencyKey'
        blockchains:
          type: array
          items:
            $ref: '#/components/schemas/Blockchain'
          description: Blockchain(s) the requested wallets will be created on.
          example:
            - ETH
          minItems: 1
        metadata:
          type: array
          description: List of metadata fields to associate with the corresponding wallet.
          items:
            $ref: '#/components/schemas/MetadataField'
    ChallengeId:
      type: string
      format: uuid
      description: Unique system generated identifier used to initiate a user challenge flow.
      example: c4d1da72-111e-4d52-bdbf-2e74a2d803d5
    SetPinResponse:
      type: object
      required:
        - challengeId
      properties:
        challengeId:
          $ref: '#/components/schemas/ChallengeId'
    Challenge:
      type: object
      required:
        - id
        - type
        - status
      properties:
        id:
          $ref: '#/components/schemas/schemas_Id'
        correlationIds:
          type: array
          items:
            type: string
          description: |
            For type CREATE_WALLET, the wallet id associated with the request. For types
            CREATE_TRANSACTION, ACCELERATE_TRANSACTION, and CANCEL_TRANSACTION, the transaction id associated with the request. 
          example:
            - 54399e5a-1bf6-4921-9559-10c1115678cd
        errorCode:
          description: Indicates the failure reason of a challenge. Only present for challenges in FAILED status.
          example: 155121
          type: integer
        errorMessage:
          description: Description of the error or error reason. Only present for challenges in FAILED status.
          example: challenge expired
          type: string
        status:
          type: string
          enum:
            - PENDING
            - IN_PROGRESS
            - COMPLETE
            - FAILED
            - EXPIRED
          example: PENDING
        type:
          type: string
          enum:
            - ''
            - INITIALIZE
            - SET_PIN
            - CHANGE_PIN
            - SET_SECURITY_QUESTIONS
            - CREATE_WALLET
            - RESTORE_PIN
            - CREATE_TRANSACTION
            - ACCELERATE_TRANSACTION
            - CANCEL_TRANSACTION
            - CONTRACT_EXECUTION
          description: The challenge type
          example: SET_PIN
    GetEndUserChallengesResponse:
      type: object
      properties:
        challenges:
          type: array
          items:
            $ref: '#/components/schemas/Challenge'
    GetEndUserChallengeByIDResponse:
      type: object
      properties:
        challenge:
          $ref: '#/components/schemas/Challenge'
    ChangePinRequest:
      type: object
      required:
        - idempotencyKey
      properties:
        idempotencyKey:
          $ref: '#/components/schemas/schemas_IdempotencyKey'
    ChangePinResponse:
      type: object
      required:
        - challengeId
      properties:
        challengeId:
          $ref: '#/components/schemas/ChallengeId'
    SetPinRequest:
      type: object
      required:
        - idempotencyKey
      properties:
        idempotencyKey:
          $ref: '#/components/schemas/schemas_IdempotencyKey'
    RestorePinRequest:
      type: object
      required:
        - idempotencyKey
      properties:
        idempotencyKey:
          $ref: '#/components/schemas/schemas_IdempotencyKey'
    RestorePinResponse:
      type: object
      required:
        - challengeId
      properties:
        challengeId:
          $ref: '#/components/schemas/ChallengeId'
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: PREFIX:ID:SECRET
      description: Circle's API Keys are formatted in the following structure "PREFIX:ID:SECRET". All three parts are requred to make a successful request.
  headers:
    XRequestId:
      description: Universally unique identifier (UUID v4) for the request. Helpful for identifying a request when communicating with Circle support.
      schema:
        type: string
        format: uuid
        example: 2adba88e-9d63-44bc-b975-9b6ae3440dde
  parameters:
    From:
      name: from
      description: From is the start time of the query, inclusive
      in: query
      schema:
        type: string
        format: date-time
        example: '2023-01-01T12:04:05Z'
    To:
      name: to
      description: To is the end time of the query, inclusive, default to current time
      in: query
      schema:
        type: string
        format: date-time
        example: '2023-01-01T12:04:05Z'
    PageBefore:
      name: pageBefore
      description: PageBefore is used to return items before the given item exclusively. SHOULD NOT be used in conjunction with pageAfter
      in: query
      schema:
        type: string
        format: uuid
    PageAfter:
      name: pageAfter
      description: PageAfter is used to return items after the given item exclusively. SHOULD NOT be used in conjunction with pageBefore
      in: query
      schema:
        type: string
        format: uuid
    PageSize:
      name: pageSize
      description: PageSize is the number of items to return
      in: query
      schema:
        type: integer
        default: 10
        maximum: 50
        minimum: 1
...
    PinStatus:
      name: pinStatus
      description: PinStatus
      in: query
      schema:
        type: string
        enum:
          - ENABLED
          - UNSET
          - LOCKED
    SecurityQuestionStatus:
      name: securityQuestionStatus
      description: SecurityQuestionStatus
      in: query
      schema:
        type: string
        enum:
          - ENABLED
          - UNSET
          - LOCKED
    XUserTokenRequired:
      name: X-User-Token
      description: User token
      required: true
      in: header
      schema:
        type: string
    Status:
      description: ExtChallenge status to query
      in: query
      name: status
      schema:
        enum:
          - PENDING
          - IN_PROGRESS
        type: string
  responses:
    Default:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      description: Error response
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'